Move runner to module `assertions`
