+++ assertions/src/commonTest/kotlin/com/github/jcornaz/kwik/assertions/AbstractRunnerTest.kt
@@ -1,7 +1,6 @@
-package com.github.jcornaz.kwik.runner
+package com.github.jcornaz.kwik.assertions
 
 import com.github.jcornaz.kwik.generator.Generator
-import com.github.jcornaz.kwik.forAll
 import kotlin.random.Random
 import kotlin.test.Test
 import kotlin.test.assertEquals
