+++ assertions/src/commonTest/kotlin/com/github/jcornaz/kwik/assertions/CustomException.kt
@@ -1,4 +1,4 @@
-package com.github.jcornaz.kwik.runner
+package com.github.jcornaz.kwik.assertions
 
 /**
  * Dummy exception to be thrown in test and make sure that the SUT don't wrap the exceptions
