+++ assertions/src/commonTest/kotlin/com/github/jcornaz/kwik/assertions/ForAll2Test.kt
@@ -1,38 +1,37 @@
-package com.github.jcornaz.kwik.runner
+package com.github.jcornaz.kwik.assertions
 
-import com.github.jcornaz.kwik.FalsifiedPropertyError
 import com.github.jcornaz.kwik.generator.Generator
-import com.github.jcornaz.kwik.forAll
 import com.github.jcornaz.kwik.withSamples
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
 import kotlin.test.assertTrue
 
-class ForAll3Test : AbstractRunnerTest() {
+class ForAll2Test : AbstractRunnerTest() {
 
     private val testGenerator1 = Generator.create { it.nextInt() }
     private val testGenerator2 = Generator.create { it.nextDouble() }
-    private val testGenerator3 = Generator.create { it.nextLong() }
 
     override fun evaluate(iterations: Int, seed: Long, invocation: () -> Boolean) {
-        forAll(testGenerator1, testGenerator2, testGenerator3, iterations, seed) { _, _, _ -> invocation() }
+        forAll(
+            testGenerator1,
+            testGenerator2,
+            iterations,
+            seed
+        ) { _, _ -> invocation() }
     }
 
     @Test
     fun evaluateSamples() {
         val ints = mutableSetOf<Int>()
         val doubles = mutableSetOf<Double>()
-        val longs = mutableSetOf<Long>()
 
         val gen1 = Generator.create { it.nextInt(0, 10) }.withSamples(42, 100)
         val gen2 = Generator.create { it.nextDouble(0.0, 10.0) }.withSamples(123.0, 678.0)
-        val gen3 = Generator.create { it.nextLong(0L, 10L) }.withSamples(-42L)
 
-        forAll(gen1, gen2, gen3) { i, d, l ->
+        forAll(gen1, gen2) { i, d ->
             ints += i
             doubles += d
-            longs += l
             true
         }
 
@@ -40,7 +39,6 @@ class ForAll3Test : AbstractRunnerTest() {
         assertTrue(100 in ints)
         assertTrue(123.0 in doubles)
         assertTrue(123.0 in doubles)
-        assertTrue(-42L in longs)
     }
 
     @Test
@@ -50,9 +48,8 @@ class ForAll3Test : AbstractRunnerTest() {
             forAll(
                 Generator.create { 42 },
                 Generator.create { -4.1 },
-                Generator.create { 100L },
                 iterations = 123, seed = 78
-            ) { _, _, _ -> ++i < 12 }
+            ) { _, _ -> ++i < 12 }
         }
 
         assertEquals(
@@ -60,7 +57,6 @@ class ForAll3Test : AbstractRunnerTest() {
                 Property falsified after 12 tests (out of 123)
                 Argument 1: 42
                 Argument 2: -4.1
-                Argument 3: 100
                 Generation seed: 78
             """.trimIndent(),
             exception.message
@@ -71,12 +67,10 @@ class ForAll3Test : AbstractRunnerTest() {
     fun evaluateForRandomValues() {
         val valuesA = mutableSetOf<Int>()
         val valuesB = mutableSetOf<Double>()
-        val valuesC = mutableSetOf<Long>()
 
-        forAll(testGenerator1, testGenerator2, testGenerator3, seed = 0L) { a, b, c ->
+        forAll(testGenerator1, testGenerator2, seed = 0L) { a, b ->
             valuesA += a
             valuesB += b
-            valuesC += c
             true
         }
 
@@ -85,18 +79,18 @@ class ForAll3Test : AbstractRunnerTest() {
 
     @Test
     fun isPredictable() {
-        val pass1 = mutableListOf<Triple<Int, Double, Long>>()
-        val pass2 = mutableListOf<Triple<Int, Double, Long>>()
+        val pass1 = mutableListOf<Pair<Int, Double>>()
+        val pass2 = mutableListOf<Pair<Int, Double>>()
 
         val seed = 123564L
 
-        forAll(testGenerator1, testGenerator2, testGenerator3, seed = seed) { a, b, c ->
-            pass1 += Triple(a, b, c)
+        forAll(testGenerator1, testGenerator2, seed = seed) { a, b ->
+            pass1 += a to b
             true
         }
 
-        forAll(testGenerator1, testGenerator2, testGenerator3, seed = seed) { a, b, c ->
-            pass2 += Triple(a, b, c)
+        forAll(testGenerator1, testGenerator2, seed = seed) { a, b ->
+            pass2 += a to b
             true
         }
 
@@ -106,8 +100,8 @@ class ForAll3Test : AbstractRunnerTest() {
     @Test
     @Suppress("USELESS_IS_CHECK")
     fun canBeCalledWithoutExplicitGenerator() {
-        forAll { a: Int, b: Long, c: Double ->
-            a is Int && b is Long && c is Double
+        forAll { a: Int, b: Long ->
+            a is Int && b is Long
         }
     }
 }
